+++ public/admin/_app/immutable/nodes/24.Bv5FHegr.js
@@ -1,0 +1,1 @@
+import{t as e}from"../chunks/C8OD1a7_.js";export{e as component};